Server deploy snapshot (lenovo, 4x A4000): py3.14/cu130 build
Captures the running config on this host so it is kept on the fork, not
discarded:
- requirements/cuda.txt: flashinfer-python/cubin commented out (no cp314
  wheel on Python 3.14); runtime sets VLLM_USE_FLASHINFER_SAMPLER=0.
- diffusion_gemma.py: pre-review TP self-conditioning fix as built into
  the current precompiled editable install (vllm ...gc6fefaf3d).

+++ vllm/model_executor/models/diffusion_gemma.py
@@ -1110,12 +1110,10 @@ class DiffusionSampler:
 
         max_num_reqs = diffusion_states.max_num_reqs
         device = diffusion_states.device
-        # int64 token ids: matches the standard Sampler output and satisfies the
-        # pipeline-parallel broadcast, which asserts sampled_token_ids is int64.
         self._sampled = torch.zeros(
             max_num_reqs,
             self.canvas_length,
-            dtype=torch.int64,
+            dtype=torch.int32,
             device=device,
         )
         self._num_sampled = torch.zeros(
